+++ src/spatial_ref/mod.rs
@@ -1,7 +1,7 @@
 mod srs;
 
-pub use srs::CoordTransform;
-pub use srs::SpatialRef;
+pub use srs::{ CoordTransform, SpatialRef, AxisOrientationType };
+pub use gdal_sys::{ OGRAxisOrientation };
 
 #[cfg(test)]
 mod tests;