+++ src/spatial_ref/tests.rs
@@ -222,3 +222,89 @@ fn axis_mapping_strategy() {
         gdal_sys::OSRAxisMappingStrategy::OAMS_TRADITIONAL_GIS_ORDER
     );
 }
+
+#[cfg(major_ge_3)]
+#[test]
+fn get_area_of_use() {
+    let spatial_ref = SpatialRef::from_epsg(4326).unwrap();
+    let area_of_use = spatial_ref.get_area_of_use().unwrap();
+    assert_eq!(area_of_use.west_lon_degree, -180.0);
+    assert_eq!(area_of_use.south_lat_degree, -90.0);
+    assert_eq!(area_of_use.east_lon_degree, 180.0);
+    assert_eq!(area_of_use.north_lat_degree, 90.0);
+}
+
+#[cfg(major_ge_3)]
+#[test]
+fn get_name() {
+    let spatial_ref = SpatialRef::from_epsg(4326).unwrap();
+    let name = spatial_ref.get_name().unwrap();
+    assert_eq!(name,"WGS 84");
+}
+
+
+#[test]
+fn get_units_epsg4326() {
+    let spatial_ref = SpatialRef::from_epsg(4326).unwrap();
+
+    let angular_units_name = spatial_ref.get_angular_units_name().unwrap();
+    assert_eq!(angular_units_name.to_lowercase(), "degree");
+    let to_radians = spatial_ref.get_angular_units();
+    assert_almost_eq(to_radians, 0.01745329);
+}
+
+#[test]
+fn get_units_epsg2154() {
+    let spatial_ref = SpatialRef::from_epsg(2154).unwrap();
+    let linear_units_name = spatial_ref.get_linear_units_name().unwrap();
+    assert_eq!(linear_units_name.to_lowercase(), "metre");
+    let to_meters = spatial_ref.get_linear_units();
+    assert_almost_eq(to_meters, 1.0);
+}
+
+#[test]
+fn predicats_epsg4326() {
+    let spatial_ref_4326 = SpatialRef::from_epsg(4326).unwrap();
+    assert!(spatial_ref_4326.is_geographic());
+    assert!(!spatial_ref_4326.is_local());
+    assert!(!spatial_ref_4326.is_projected());
+    assert!(!spatial_ref_4326.is_compound());
+    assert!(!spatial_ref_4326.is_geocentric());
+    assert!(!spatial_ref_4326.is_vertical());
+
+    #[cfg(all(major_ge_3, minor_ge_1))]
+    assert!(!spatial_ref_4326.is_derived_geographic());
+}
+
+#[test]
+fn predicats_epsg2154() {
+    let spatial_ref_2154 = SpatialRef::from_epsg(2154).unwrap();
+    assert!(!spatial_ref_2154.is_geographic());
+    assert!(!spatial_ref_2154.is_local());
+    assert!(spatial_ref_2154.is_projected());
+    assert!(!spatial_ref_2154.is_compound());
+    assert!(!spatial_ref_2154.is_geocentric());
+
+    #[cfg(all(major_ge_3, minor_ge_1))]
+    assert!(!spatial_ref_2154.is_derived_geographic());
+} 
+
+//XXX Gdal 2 implementation is partial
+#[cfg(major_ge_3)] 
+#[test]
+fn crs_axis() {
+    let spatial_ref = SpatialRef::from_epsg(4326).unwrap();
+
+    #[cfg(all(major_ge_3, minor_ge_1))]
+    assert_eq!(spatial_ref.get_axes_count(),2);
+    
+    let orientation = spatial_ref.get_axis_orientation("GEOGCS",0);
+    assert_eq!(orientation,gdal_sys::OGRAxisOrientation::OAO_North);
+
+    assert!(spatial_ref.get_axis_name("GEOGCS",0).is_some());
+    assert!(spatial_ref.get_axis_name("DO_NO_EXISTS",0).is_none());
+
+    let orientation = spatial_ref.get_axis_orientation("DO_NO_EXISTS",0);
+    assert_eq!(orientation, gdal_sys::OGRAxisOrientation::OAO_Other);
+}
+